shifted from execute_ssh_command to get_miner_data, still debugging
